Synchronize the option haggling phase
Currently the protocol permits options to be sent even when there
are outstanding replies, and for the server to process those
options out of order. This causes a number of issues.

If option replies remain outstanding when NBD_OPT_STARTTLS is issued
then those replies may arrive unencrypted when NBD_OPT_STARTTLS
is processed. Similarly options sent after NBD_OPT_STARTTLS is
issued (but not replied to) may be sent unencrypted but are
expected to be encrypted. This is difficult to track server side.
This needs fixing in any case, and it's better to fix it for
all options.

Few (if any) known clients issue options without waiting for replies,
and few (if any) servers do anything other than process replies.
So the ability to send options without waiting for replies gives
few benefits. It also complicates the specification, and (as
it is so little tested) is a ripe field for subtle bugs where the
protocol authors assume that a reply to a given option has been
received before the client sends another option.

Without this serialisation, it is not possible to effectively
enforce things like "The client MUST NOT send NBD_OPT_X if
the server has previously replied NBD_REP_Y" because the
client can send NBD_OPT_X before receiving the reply.

Signed-off-by: Alex Bligh <alex@alex.org.uk>