doc: Specific error if NBD_OPT_GO refused without NBD_OPT_BLOCK_SIZE
While we documented that a server could refuse a client that
did not mention that it would handle non-default block sizes,
we didn't mention what error to use.  For NBD_OPT_EXPORT_NAME,
we have no choice (a hard disconnect is the only way possible).
But for NBD_OPT_GO (and similarly for NBD_OPT_INFO), the
existing NBD_REP_ERR_POLICY and NBD_REP_ERR_INVALID feel rather
weak, so it's better to invent a specific
NBD_REP_ERR_BLOCK_SIZE_REQD.

Suggested-by: Alex Bligh <alex@alex.org.uk>
Signed-off-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Alex Bligh <alex@alex.org.uk>