protocol: Ignore rest of option when replying with error
We must consume the payload of any option received, whether or
not we can answer the option successfully, in order to be in
sync for reading the next option.  This is particularly true
once clients start using NBD_OPT_GO, as long as we don't happen
to service that option.

Signed-off-by: Eric Blake <eblake@redhat.com>