Improve backwards compatibility.
While writing the protocol document, we decided that a server should
always send a payload, even when an error occurred. Only it turned out
that this was not in line with what existing implementations did.

Since
- It seems to make more sense to not send invalid data to a client when
  we know it to be invalid, and
- this is what current implementations do at any rate,

change the spec to say that servers MUST NOT send payload on read error,
and change the implementation to match.