Stop generating nbd-client configuration
Ten years ago, generating configuration from debconf sounded like a good
idea. That was wrong, and I should never have done it. But hey, it
existed, it seemed to work, and maybe someone was using it, right?

Yeah, well, not anymore. Rewriting all of that to support
/etc/nbdtab would have been way too complicated, and there's not really
any reason why we would want to, so use the migration from
/etc/nbd-client to /etc/nbdtab as an excuse to junk all that.

Do add code to do a one-time migration from nbd-client to nbdtab,
though.