Prepare for allowing STARTTLS
Replace every use of readit() or writeit() on a socket by an abstracting
socket_read()/socket_write() function instead.

This socket_read()/socket_write() function will call
gnutls_record_recv()/gnutls_record_send() if the client is doing TLS,
which we signal by the non-NULL-ness of the new .tls_session member of
the CLIENT struct.

After that, once we add STARTTLS negotiation, we should be able to do
TLS properly.

Signed-off-by: Wouter Verhelst <w@uter.be>