Refactor negotiation a bit.
Currently, we did negotiation by way of direct read() and write() calls.
This has two problems:
- The error handling in the implementation is not perfect; it is
  possible to have short reads, which would result in the server
  assuming a buffer has a given size, when in reality it does not.
- Having direct read() and write() calls makes it impossible to abstract
  away whether we're using TLS or not.

The easy fix is to use the new socket_read() and socket_write() calls,
but those require a CLIENT* to be available before they can be used.
Rather than allocating it in handle_export_name(), we now allocate it in
negotiate(), passing the pointer to the export name handler, and
removing its allocation.

Signed-off-by: Wouter Verhelst <w@uter.be>