Properly handle maxconnections
The "maxconnections" setting considers connections to all exports,
rather than "just" the one where the "maxconnections" setting is set.
This is unexpected and not useful.

Fix by creating a socketpair over which the child asks the parent
process whether a connection to the given export is allowed, rather than
doing the check in the client process.

Closes github #35

Signed-off-by: Wouter Verhelst <w@uter.be>