Make nbd-client -l work again
We don't need to open an nbd device if all we're going to ask for is the
list of exports.

Also, nbd-client -l is a valid thing to run even if you don't have
access to /dev/nbd* (which the test suite does, for instance), so don't
even try to open the device node unless we really need to.

Fixes: 4e08f11ab9996d0c4a8c52d5acecf8b33890825f
Signed-off-by: Wouter Verhelst <w@uter.be>