Remove hostname confusion
Currently, there is a "hostname" argument (which specifies the server to
connect to, and which is not an option argument) and a '-hostname'
argument (which takes the hostname for SNI connectivity in the TLS
context). The two are only tangentially related, and it could be
confusing to have two "hostname" arguments.

Rename the TLS-related one to tlshostname to clarify.

Signed-off-by: Wouter Verhelst <w@uter.be>