Fix shell syntaxis
The "test" command takes one or more arguments, which are then evaluated
in the shell before processing. An assignment results in an exit state,
which as far as test is concerned, is fine enough.

Since we wrote "x$HAVE_GNUTLS=1", that would get expanded to "x0=1",
meaning, we were assigning the value 1 to the variable "x0", which meant
that "test" was seeing one argument rather than the three that we wanted
to give it, and so it was happy.

Fix by adding the required spaces to make test see the equals sign and
perform the comparison.

Shell is a horrible language.

Signed-off-by: Wouter Verhelst <w@uter.be>