Fix WRITE_ZEROES/STARTTLS combination
We merged the WRITE_ZEROES support from a branch, where it was written
before STARTTLS was implemented, without properly adapting it to the way
things are now done in the light of STARTTLS, resulting in the
WRITE_ZEROES function sending its reply header in the clear,
unconditionally.

Fix by using the correct function instead.