Don't seek() in files without holding a lock.
The introduction of multithread handling of requests failed to consider
what happens with copyonwrite enabled. We would seek, then read (or
write), a file.

If we have copyonwrite enabled, this fails badly in that two threads can
race against eachother for the current file pointer. Instead, we should
use pread() or pwrite() to read from offset without relying on the
current file offset, as explained in #43.

Remove the myseek() function to ensure we don't make use of it anymore,
and fix up all callers.

The only exception where we can still use lseek() safely is in the
open_treefile() function, since there we hold a mutex and hence we can't
race against other threads.

TODO: test whether this fixes the problem.

Signed-off-by: Wouter Verhelst <w@uter.be>