Use AM_CONDITIONAL rather than AC_SUBST
AC_SUBST'ing the nbd-client program worked when we didn't need to
compile the client twice. Now that we do, this makes us try to compile
the minimal (i.e., non-TLS) client even if we're on a system where the
client can't be compiled. This is obviously wrong.

Fix it by using AM_CONDITIONAL a second time, and nesting the whole
GnuTLS conditional below that.

Signed-off-by: Wouter Verhelst <w@uter.be>