Fix hostname:/path support
LTSP uses NFS-like export names, and expects a "/" in the name to be part of an
export name, rather than it being the delimiter between a port and an export
name.

Since NFS-like export names don't have a port number, we can
special-case the ":/" format, and use the : as a delimiter then.