Swap options and option name
When trying to reply to a client with certain information, it's useful
if we don't need to read the whole reply but can handle it by reading
one option, handling it, reading the next one, etc etc. This requires
that the server is told which export to deal with before getting the
option names, however.

As specified, that isn't the case, so this would require the server to
read in the whole request before it can start processing it.

Swap options around to make handling the INFO/GO messages easier to do.

Signed-off-by: Wouter Verhelst <w@uter.be>
Acked-By: Eric Blake <eblake@redhat.com>