Centralize what needs to be done before exporting
Previously, we ran everything that needed to be done before exporting
for a particular client (e.g., checking access rights, opening files,
running scripts, ...) in various locations all over the place. This
works, but it assumes NBD_OPT_EXPORT_NAME usage only, because it sends
the sizes and flags etc unconditionally after we've left the
negotiation() function.

This works for NBD_OPT_EXPORT_NAME, but not for NBD_OPT_GO, which needs
to send a negotiation acknowledgement before formally closing the
negotiation phase. As such, rearrange things so that they're all in one
function, commit_client, which does all the setup, and call that
function before returning from the negotiation() function.

This new function wants to call a number of other functions that need to
be moved up in the nbd-server file; but we'll do that after this commit
so as to not muddle the diff with irrelevant fluff. Add forward
declarations to them for now.

Signed-off-by: Wouter Verhelst <w@uter.be>