Turn various fatal errors into return values
Previously, we would err() out of a function whenever things went south.
That's fine for NBD_OPT_EXPORT_NAME, but not for NBD_OPT_GO, which needs
to deal with things differently.

Turn each and every one of those err() calls into an err_nonfatal, and
change the return value of the function.

While at it, harmonize the function return signatures into booleans
everywhere.

Signed-off-by: Wouter Verhelst <w@uter.be>