Regenerate certificate
Although the configuration file specified a validity of 3650 days (i.e.,
almost 10 years), the actual certificate in this directory was generated
with a certificate validity of 1 year. It expired on april 19th, 2017
(i.e., 5 days ago), causing the test suite to fail from now on.

Fix by regenerating the certificate (not the keys) of the certificate
with the actual configuration in this directory.

Signed-off-by: Wouter Verhelst <w@uter.be>