Don't forget to read the length of the request
We read out the NBD_OPT_INFO packet's length as the length of the name
(which follows it immediately), rather than the length of the packet.

Add a second read to fix that.

While at it, add a sanity check for the length of both, and send back an
NBD_REP_ERR_INVALID if it fails.

Signed-off-by: Wouter Verhelst <w@uter.be>