implemented waiting for file
If we expect an export file to appear and know its size, we can
temporarily save writes into diff file. When the file appears
we commit all writes into it and switch writes to export file.

This mode can be useful when live migrating VM between hypervisors:
We start mirroring writes from source HV to destination despite
the destination storage doesn't exist yet. Then we can start
copying the source storage (ZFS, LVM, ...) to destination and
when it appears on destination, cached writes are commited to it.