Fix double close
We had two code paths exiting the negotiate() function in error; one
which would close the network, and one which wouldn't.

This is not an ideal situation. Since closing the network socket is
something we'll always need to do anyway, it makes more sense to do it
in the negotiate function rather than in the calling function.

Coverity CID#1437710

Signed-off-by: Wouter Verhelst <w@uter.be>