Move the close() call back to where it belongs
In trying to write a fix for CID#1437710, we missed the fact that the
close in the handle_modern_connection() call is necessary for more than
just the failure return of negotiate(). As a result, we now leak a
resource in some cases.

Revert the removal of the close() call, and do it in
handle_modern_connection() anyway.

Coverity CID#1437712

Signed-off-by: Wouter Verhelst <w@uter.be>