Ensure that i is not used uninitialized
Currently, multifile and treefile are mutually exclusive (so this
message cannot be produced in that case), and the only effect this will
have is that the log message will contain a garbage integer value, but
it's not nice.

Ensure i is initialized to 0, so that at least we say that there are no
files in use, in that case.

Coverity CID#1264432

Signed-off-by: Wouter Verhelst <w@uter.be>