doc: Explicit mention that 0-length transactions are unspecified
We don't want to forbid 0-length transactions between a client
and server that have some special semantics for it, but we also
don't want to define any particular semantics.  Best is to just
document that portable clients should not attempt it, and that
servers should tolerate clients that do it anyway (whether by a
successful no-op or by an error code, we don't care, as long as
the connection to the client is not killed).

While in the area, mention that a trim request may only cause a
trim action on a subset of the request.

Signed-off-by: Eric Blake <eblake@redhat.com>