doc: Update reserved values for pending extensions
The current state of the BLOCK_STATUS extension proposal no longer
consumes an NBD_FLAG_ bit, but does add a couple of NBD_OPT_ commands
and a new NBD_REP_ reply during handshake, as well as reserving a
new NBD_CMD_FLAG_ bit during transmission.

While at it, fix inconsistent top-level bullets as used in the list
of NBD_OPT_ and NBD_REP_.

The RESIZE extension proposal does not have to skip a bit number in
the NBD_FLAG_ namespace.

Signed-off-by: Eric Blake <eblake@redhat.com>