doc: Clarify NBD_CMD_READ payload on error
Commit e6c6fb39 clarified that a payload for the reply to NBD_CMD_READ
must not be sent if the error field was set (which matches long-time
existing qemu-nbd behavior; nbd-server was switched to that behavior
at the time; and then just this week nbdkit was also patched to comply).
But it missed another spot in the document, which if read in isolation,
results in the wrong behavior.

Signed-off-by: Eric Blake <eblake@redhat.com>