server: Consolidate request validation
We had a couple of issues when dealing with bogus requests from
malicious clients, when compared to what we document in the spec
(fortunately, the client is violating protocol by sending these
requests which leaves us a bit more flexibility on our response;
and as we are not crashing, this a quality-of-implementation
rather than a security problem):
- NBD_CMD_TRIM failed with EINVAL instead of EPERM on a read-only image
- range validation for several commands was implicit based on syscall
failures, rather than being filtered out up front (which means we
are dependent on the right errno value, or even risk succeeding at
doing something out of range, although I did not audit all paths)
- At least in the case of NBD_CMD_TRIM, a req->from near 2^64 plus
req->len could overflow into something smaller than client->exportsize
- Duplicating checks everywhere is prone to bitrot

It's easier to consolidate validation up front, so we don't have
multiple copies of range checks, so we don't have to rely on
syscalls catching odd ranges, and so that we are more in line
with what the spec documents.  This in turn requires updating
the testsuite for trim, now that the validation is done in a
different place.

Signed-off-by: Eric Blake <eblake@redhat.com>