Stop using nonportable flags of readlink
"readlink -f" is a GNU extension. On OSX (where we try to run the test
suite, too, courtesy of travis CI) it does not exist.

Not sure why we ever added it in the first place (symlinks work, you
know) so just drop it entirely; there is no need to canonicalize things
here.