nbd-client: don't set unset (zero) timeout when configuring via netlink
When we are operating as a daemon we ignore unset (zero) timeout in
set_timeout() and don't set it in kernel.
However, when operating in the netlink mode netlink_configure() will add
NBD_ATTR_TIMEOUT attribute unconditionally.

This means that in the default case of timeout being unset it was being set
to zero in the kernel which resulted in it triggering almost immediately
upon any I/O request.

Fix this by not appending the timeout attribute in netlink_configure() when
the timeout is unset.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>