Don't listen on TCP if we are using a UNIX socket.
Currently, `nbd-server` will open a listening TCP socket even if it's
been asked to use a UNIX domain socket.  This is inconsistent with
typical behavior of applications which can use UNIX sockets, and may
present a security risk on some systems.

This commit modifies this behavior so that use of a UNIX socket and
other types of socket is mutually exclusive.  You can only listen on a
regular (TCP or SDP) socket or a UNIX socket, but not both, with the
prescense of the `unixsock` option taking precendence over any other
listener configuration.

This has a small potential to break some existing installations, though
they are arguably depending on broken behavior to begin with.

The manual page is also updated to reflect this change.

Signed-off-by: Austin S. Hemmelgarn <ahferroin7@gmail.com>