Convert SERVER to a refcounted value
We want to eventually be able to dynamically add and/or remove server
configurations. In order to be able to do that, it's better to refcount
the loaded configuration, so that we can then dereference it when it has
changed, without overwriting things.

This requires us to change the servers GArray from one holding objects
to one holding pointers to them, so do so.