doc: Clarify use of EOVERFLOW
The error EOVERFLOW was added as part of the NBD_CMD_FLAG_DF flag for
structured reads. Since NBD errors are not necessarily 1:1 with local
errno values, it is conceivable that a client may not be expecting to
receive this value, so it is better to document that a server should
not send it in situations that older clients were not expecting.

Signed-off-by: Eric Blake <eblake@redhat.com>