doc: Tweak naming of NBD errors
It is intentional that we send NBD-specific error values over the
wire, which (due to history) happen to match particular Linux errno
values, but which may not match on other systems.  To make it more
obvious that a mapping is involved (both for the server converting
local errno values into wire values, and in clients converting wire
values back into any human messages via strerror() or similar), update
the protocol specification to refer to NBD errors with an NBD_E
prefix.

Signed-off-by: Eric Blake <eblake@redhat.com>