Segfault in nbd-server.c due to calling freeaddrinfo on NULL
In set_peername, when netaddr.ss_family == AF_UNIX. ai is never initialized.

On line 1733, freeaddrinfo is called unconditionally on ai, which might be NULL
in some cases.

The musl implementation of getaddrinfo started to segfault when you pass it
NULL. This caused some of the tests to fail on Alpine Linux, namely simple_test
inetd and unix.

This patch was enough to fix the issue