docs: Recommend TCP_NODELAY (disabling Nagle)
This text from https://en.wikipedia.org/wiki/Nagle%27s_algorithm
is telling:

"In any request-response application protocols where request data can
be larger than a packet, this can artificially impose a few hundred
milliseconds latency between the requester and the responder, even if
the requester has properly buffered the request data. Nagle's
algorithm should be disabled by the requester in this case. If the
response data can be larger than a packet, the responder should also
disable Nagle's algorithm so the requester can promptly receive the
whole response."

And we hit it in real life with libnbd/nbdkit - more than an order of
magnitude in slowdown results when trying to saturate a connection
with as many parallel requests as possible when Nagle's algorithm
introduces artificial delay waiting for the ACKs:
https://www.redhat.com/archives/libguestfs/2019-May/msg00237.html

Signed-off-by: Eric Blake <eblake@redhat.com>